Updated AccessToken's hasExpired
Correct me if I am wrong but checking the expired time does not work if we use the right now time.
We want to have some time to use the token after we check that it is still valid right? For example when you make a couple requests to an API with a part of the response of the previous request.

I know that the chance of this happening is really small but still.